requirements.txt
exchangelib==5.4.0
pandas>=2.0.0
python-dotenv==1.0.0
schedule==1.2.0
tqdm==4.66.1
colorama==0.4.6
flask==2.3.3
flask-sqlalchemy==3.0.5
werkzeug==2.3.7 